mcp_servers/database_mcp_server/requirements.txt
fastmcp
SQLAlchemy
aiomysql
python-dotenv 
greenlet